services/notification/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
celery==5.3.4
redis==5.0.1
pika==1.3.2
django-filter==23.5
requests==2.31.0
django-cors-headers==4.3.1
